Init Commit: C++ quick sort implementation
(this is sample code from an assignment at UNLV for CS 302 during the
Fall 2013 Semester). This C++ program was created to test and compare
the quick sort algorithm. As the size of a list to be sorted grows, the
quick sort algorithm dramatically outperforms it’s selection sort
counterpart. My final report PDF shows these results.